requirements.txt
appdirs==1.4.4
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
decorator==5.1.1
Flask==2.3.3
fredapi==0.5.1
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.24.4
pandas==1.3.4
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
tabulate==0.9.0
urllib3==2.0.4
wbdata==0.3.0
Werkzeug==2.3.7
zipp==3.16.2
